Move Instruction list (clickable instruction works again)
